+++ src/internal/connector/exchange/attendees.go
@@ -1,4 +1,4 @@
-package support
+package exchange
 
 import (
 	"fmt"
