+++ src/pkg/services/m365/api/contacts.go
@@ -38,13 +38,17 @@ type Contacts struct {
 // If successful, returns the created folder object.
 func (c Contacts) CreateContactFolder(
 	ctx context.Context,
-	user, folderName string,
+	userID, containerName string,
 ) (models.ContactFolderable, error) {
-	requestBody := models.NewContactFolder()
-	temp := folderName
-	requestBody.SetDisplayName(&temp)
+	body := models.NewContactFolder()
+	body.SetDisplayName(ptr.To(containerName))
 
-	mdl, err := c.Stable.Client().Users().ByUserId(user).ContactFolders().Post(ctx, requestBody, nil)
+	mdl, err := c.Stable.
+		Client().
+		Users().
+		ByUserId(userID).
+		ContactFolders().
+		Post(ctx, body, nil)
 	if err != nil {
 		return nil, graph.Wrap(ctx, err, "creating contact folder")
 	}
@@ -55,7 +59,7 @@ func (c Contacts) CreateContactFolder(
 // DeleteContainer deletes the ContactFolder associated with the M365 ID if permissions are valid.
 func (c Contacts) DeleteContainer(
 	ctx context.Context,
-	user, folderID string,
+	userID, containerID string,
 ) error {
 	// deletes require unique http clients
 	// https://github.com/alcionai/corso/issues/2707
@@ -64,7 +68,13 @@ func (c Contacts) DeleteContainer(
 		return graph.Stack(ctx, err)
 	}
 
-	err = srv.Client().Users().ByUserId(user).ContactFolders().ByContactFolderId(folderID).Delete(ctx, nil)
+	err = srv.
+		Client().
+		Users().
+		ByUserId(userID).
+		ContactFolders().
+		ByContactFolderId(containerID).
+		Delete(ctx, nil)
 	if err != nil {
 		return graph.Stack(ctx, err)
 	}
@@ -79,18 +89,14 @@ func (c Contacts) GetFolder(
 	ctx context.Context,
 	userID, containerID string,
 ) (models.ContactFolderable, error) {
-	service, err := c.Service()
-	if err != nil {
-		return nil, graph.Stack(ctx, err)
-	}
-
 	config := &users.ItemContactFoldersContactFolderItemRequestBuilderGetRequestConfiguration{
 		QueryParameters: &users.ItemContactFoldersContactFolderItemRequestBuilderGetQueryParameters{
 			Select: idAnd(displayName, parentFolderID),
 		},
 	}
 
-	resp, err := service.Client().
+	resp, err := c.Stable.
+		Client().
 		Users().
 		ByUserId(userID).
 		ContactFolders().
@@ -106,9 +112,9 @@ func (c Contacts) GetFolder(
 // interface-compliant wrapper of GetFolder
 func (c Contacts) GetContainerByID(
 	ctx context.Context,
-	userID, dirID string,
+	userID, containerID string,
 ) (graph.Container, error) {
-	return c.GetFolder(ctx, userID, dirID)
+	return c.GetFolder(ctx, userID, containerID)
 }
 
 func (c Contacts) PatchFolder(
@@ -116,12 +122,8 @@ func (c Contacts) PatchFolder(
 	userID, containerID string,
 	body models.ContactFolderable,
 ) error {
-	service, err := c.Service()
-	if err != nil {
-		return graph.Stack(ctx, err)
-	}
-
-	_, err = service.Client().
+	_, err := c.Stable.
+		Client().
 		Users().
 		ByUserId(userID).
 		ContactFolders().
@@ -145,15 +147,10 @@ func (c Contacts) PatchFolder(
 // not contain historical data.
 func (c Contacts) EnumerateContainers(
 	ctx context.Context,
-	userID, baseDirID string,
+	userID, baseContainerID string,
 	fn func(graph.CachedContainer) error,
 	errs *fault.Bus,
 ) error {
-	service, err := c.Service()
-	if err != nil {
-		return graph.Stack(ctx, err)
-	}
-
 	config := &users.ItemContactFoldersItemChildFoldersRequestBuilderGetRequestConfiguration{
 		QueryParameters: &users.ItemContactFoldersItemChildFoldersRequestBuilderGetQueryParameters{
 			Select: idAnd(displayName, parentFolderID),
@@ -161,11 +158,12 @@ func (c Contacts) EnumerateContainers(
 	}
 
 	el := errs.Local()
-	builder := service.Client().
+	builder := c.Stable.
+		Client().
 		Users().
 		ByUserId(userID).
 		ContactFolders().
-		ByContactFolderId(baseDirID).
+		ByContactFolderId(baseContainerID).
 		ChildFolders()
 
 	for {
@@ -205,7 +203,7 @@ func (c Contacts) EnumerateContainers(
 			break
 		}
 
-		builder = users.NewItemContactFoldersItemChildFoldersRequestBuilder(link, service.Adapter())
+		builder = users.NewItemContactFoldersItemChildFoldersRequestBuilder(link, c.Stable.Adapter())
 	}
 
 	return el.Failure()
@@ -218,7 +216,7 @@ func (c Contacts) EnumerateContainers(
 // GetItem retrieves a Contactable item.
 func (c Contacts) GetItem(
 	ctx context.Context,
-	user, itemID string,
+	userID, itemID string,
 	immutableIDs bool,
 	_ *fault.Bus, // no attachments to iterate over, so this goes unused
 ) (serialization.Parsable, *details.ExchangeInfo, error) {
@@ -226,7 +224,13 @@ func (c Contacts) GetItem(
 		Headers: newPreferHeaders(preferImmutableIDs(immutableIDs)),
 	}
 
-	cont, err := c.Stable.Client().Users().ByUserId(user).Contacts().ByContactId(itemID).Get(ctx, options)
+	cont, err := c.Stable.
+		Client().
+		Users().
+		ByUserId(userID).
+		Contacts().
+		ByContactId(itemID).
+		Get(ctx, options)
 	if err != nil {
 		return nil, nil, graph.Stack(ctx, err)
 	}
@@ -239,12 +243,8 @@ func (c Contacts) PostItem(
 	userID, containerID string,
 	body models.Contactable,
 ) (models.Contactable, error) {
-	service, err := c.Service()
-	if err != nil {
-		return nil, graph.Stack(ctx, err)
-	}
-
-	itm, err := service.Client().
+	itm, err := c.Stable.
+		Client().
 		Users().
 		ByUserId(userID).
 		ContactFolders().
@@ -264,12 +264,13 @@ func (c Contacts) DeleteItem(
 ) error {
 	// deletes require unique http clients
 	// https://github.com/alcionai/corso/issues/2707
-	service, err := c.Service()
+	srv, err := c.Service()
 	if err != nil {
 		return graph.Stack(ctx, err)
 	}
 
-	err = service.Client().
+	err = srv.
+		Client().
 		Users().
 		ByUserId(userID).
 		Contacts().
@@ -297,7 +298,7 @@ type contactPager struct {
 func NewContactPager(
 	ctx context.Context,
 	gs graph.Servicer,
-	user, directoryID string,
+	userID, containerID string,
 	immutableIDs bool,
 ) itemPager {
 	config := &users.ItemContactFoldersItemContactsRequestBuilderGetRequestConfiguration{
@@ -307,11 +308,12 @@ func NewContactPager(
 		Headers: newPreferHeaders(preferPageSize(maxNonDeltaPageSize), preferImmutableIDs(immutableIDs)),
 	}
 
-	builder := gs.Client().
+	builder := gs.
+		Client().
 		Users().
-		ByUserId(user).
+		ByUserId(userID).
 		ContactFolders().
-		ByContactFolderId(directoryID).
+		ByContactFolderId(containerID).
 		Contacts()
 
 	return &contactPager{gs, builder, config}
@@ -345,8 +347,8 @@ var _ itemPager = &contactDeltaPager{}
 
 type contactDeltaPager struct {
 	gs          graph.Servicer
-	user        string
-	directoryID string
+	userID      string
+	containerID string
 	builder     *users.ItemContactFoldersItemContactsDeltaRequestBuilder
 	options     *users.ItemContactFoldersItemContactsDeltaRequestBuilderGetRequestConfiguration
 }
@@ -354,18 +356,17 @@ type contactDeltaPager struct {
 func getContactDeltaBuilder(
 	ctx context.Context,
 	gs graph.Servicer,
-	user string,
-	directoryID string,
+	userID, containerID string,
 	options *users.ItemContactFoldersItemContactsDeltaRequestBuilderGetRequestConfiguration,
 ) *users.ItemContactFoldersItemContactsDeltaRequestBuilder {
-	builder := gs.Client().Users().ByUserId(user).ContactFolders().ByContactFolderId(directoryID).Contacts().Delta()
+	builder := gs.Client().Users().ByUserId(userID).ContactFolders().ByContactFolderId(containerID).Contacts().Delta()
 	return builder
 }
 
 func NewContactDeltaPager(
 	ctx context.Context,
 	gs graph.Servicer,
-	user, directoryID, deltaURL string,
+	userID, containerID, oldDelta string,
 	immutableIDs bool,
 ) itemPager {
 	options := &users.ItemContactFoldersItemContactsDeltaRequestBuilderGetRequestConfiguration{
@@ -376,13 +377,13 @@ func NewContactDeltaPager(
 	}
 
 	var builder *users.ItemContactFoldersItemContactsDeltaRequestBuilder
-	if deltaURL != "" {
-		builder = users.NewItemContactFoldersItemContactsDeltaRequestBuilder(deltaURL, gs.Adapter())
+	if oldDelta != "" {
+		builder = users.NewItemContactFoldersItemContactsDeltaRequestBuilder(oldDelta, gs.Adapter())
 	} else {
-		builder = getContactDeltaBuilder(ctx, gs, user, directoryID, options)
+		builder = getContactDeltaBuilder(ctx, gs, userID, containerID, options)
 	}
 
-	return &contactDeltaPager{gs, user, directoryID, builder, options}
+	return &contactDeltaPager{gs, userID, containerID, builder, options}
 }
 
 func (p *contactDeltaPager) getPage(ctx context.Context) (DeltaPageLinker, error) {
@@ -399,7 +400,7 @@ func (p *contactDeltaPager) setNext(nextLink string) {
 }
 
 func (p *contactDeltaPager) reset(ctx context.Context) {
-	p.builder = getContactDeltaBuilder(ctx, p.gs, p.user, p.directoryID, p.options)
+	p.builder = getContactDeltaBuilder(ctx, p.gs, p.userID, p.containerID, p.options)
 }
 
 func (p *contactDeltaPager) valuesIn(pl PageLinker) ([]getIDAndAddtler, error) {
@@ -408,35 +409,38 @@ func (p *contactDeltaPager) valuesIn(pl PageLinker) ([]getIDAndAddtler, error) {
 
 func (c Contacts) GetAddedAndRemovedItemIDs(
 	ctx context.Context,
-	user, directoryID, oldDelta string,
+	userID, containerID, oldDelta string,
 	immutableIDs bool,
 	canMakeDeltaQueries bool,
 ) ([]string, []string, DeltaUpdate, error) {
-	service, err := c.Service()
-	if err != nil {
-		return nil, nil, DeltaUpdate{}, graph.Stack(ctx, err)
-	}
-
 	ctx = clues.Add(
 		ctx,
 		"category", selectors.ExchangeContact,
-		"container_id", directoryID)
+		"container_id", containerID)
 
-	pager := NewContactPager(ctx, service, user, directoryID, immutableIDs)
-	deltaPager := NewContactDeltaPager(ctx, service, user, directoryID, oldDelta, immutableIDs)
+	pager := NewContactPager(ctx, c.Stable, userID, containerID, immutableIDs)
+	deltaPager := NewContactDeltaPager(ctx, c.Stable, userID, containerID, oldDelta, immutableIDs)
 
-	return getAddedAndRemovedItemIDs(ctx, service, pager, deltaPager, oldDelta, canMakeDeltaQueries)
+	return getAddedAndRemovedItemIDs(ctx, c.Stable, pager, deltaPager, oldDelta, canMakeDeltaQueries)
 }
 
 // ---------------------------------------------------------------------------
 // Serialization
 // ---------------------------------------------------------------------------
 
-// Serialize rserializes the item into a byte slice.
+func BytesToContactable(bytes []byte) (models.Contactable, error) {
+	v, err := createFromBytes(bytes, models.CreateContactFromDiscriminatorValue)
+	if err != nil {
+		return nil, clues.Wrap(err, "deserializing bytes to contact")
+	}
+
+	return v.(models.Contactable), nil
+}
+
 func (c Contacts) Serialize(
 	ctx context.Context,
 	item serialization.Parsable,
-	user, itemID string,
+	userID, itemID string,
 ) ([]byte, error) {
 	contact, ok := item.(models.Contactable)
 	if !ok {
@@ -444,15 +448,11 @@ func (c Contacts) Serialize(
 	}
 
 	ctx = clues.Add(ctx, "item_id", ptr.Val(contact.GetId()))
-
-	var (
-		err    error
-		writer = kjson.NewJsonSerializationWriter()
-	)
+	writer := kjson.NewJsonSerializationWriter()
 
 	defer writer.Close()
 
-	if err = writer.WriteObjectValue("", contact); err != nil {
+	if err := writer.WriteObjectValue("", contact); err != nil {
 		return nil, graph.Stack(ctx, err)
 	}
